+++ src/render/mod.rs
@@ -17,10 +17,24 @@ use crate::prelude::Collider;
 #[derive(Copy, Clone, Component, PartialEq, Debug)]
 pub struct ColliderDebugColor(pub Hsla);
 
+/// Overrides the global [`DebugRenderContext`] for a single collider.
+#[derive(Copy, Clone, Reflect, Component, Eq, PartialEq, Default, Debug)]
+pub enum ColliderDebug {
+    /// Always render the debug gizmos for this collider, regardless of global config.
+    #[default]
+    AlwaysRender,
+    /// Never render the debug gizmos for this collider, regardless of global config.
+    NeverRender,
+}
+
 /// Plugin rensponsible for rendering (using lines) what Rapier "sees" when performing
 /// its physics simulation. This is typically useful to check proper
 /// alignment between colliders and your own visual assets.
 pub struct RapierDebugRenderPlugin {
+    /// Whether to show debug gizmos for all colliders.
+    ///
+    /// Can be overridden for individual colliders by adding a [`ColliderDebug`] component.
+    pub default_collider_debug: ColliderDebug,
     /// Is the debug-rendering enabled?
     pub enabled: bool,
     /// Control some aspects of the render coloring.
@@ -36,6 +50,7 @@ impl Default for RapierDebugRenderPlugin {
     fn default() -> Self {
         Self {
             enabled: true,
+            default_collider_debug: ColliderDebug::AlwaysRender,
             style: DebugRenderStyle {
                 rigid_body_axes_length: 20.0,
                 ..Default::default()
@@ -47,6 +62,7 @@ impl Default for RapierDebugRenderPlugin {
     fn default() -> Self {
         Self {
             enabled: true,
+            default_collider_debug: ColliderDebug::AlwaysRender,
             style: DebugRenderStyle::default(),
             mode: DebugRenderMode::default(),
         }
@@ -67,6 +83,10 @@ impl RapierDebugRenderPlugin {
 pub struct DebugRenderContext {
     /// Is the debug-rendering currently enabled?
     pub enabled: bool,
+    /// Whether to show debug gizmos for all colliders.
+    ///
+    /// Can be overriden for individual colliders by adding a [`ColliderDebug`] component.
+    pub default_collider_debug: ColliderDebug,
     /// Pipeline responsible for rendering. Access `pipeline.mode` and `pipeline.style`
     /// to modify the set of rendered elements, and modify the default coloring rules.
     #[reflect(ignore)]
@@ -77,6 +97,7 @@ impl Default for DebugRenderContext {
     fn default() -> Self {
         Self {
             enabled: true,
+            default_collider_debug: ColliderDebug::AlwaysRender,
             pipeline: DebugRenderPipeline::default(),
         }
     }
@@ -85,9 +106,11 @@ impl Default for DebugRenderContext {
 impl Plugin for RapierDebugRenderPlugin {
     fn build(&self, app: &mut App) {
         app.register_type::<DebugRenderContext>();
+        app.register_type::<ColliderDebug>();
 
         app.insert_resource(DebugRenderContext {
             enabled: self.enabled,
+            default_collider_debug: self.default_collider_debug,
             pipeline: DebugRenderPipeline::new(self.style, self.mode),
         })
         .add_systems(
@@ -99,6 +122,8 @@ impl Plugin for RapierDebugRenderPlugin {
 
 struct BevyLinesRenderBackend<'world, 'state, 'a, 'b> {
     custom_colors: Query<'world, 'state, &'a ColliderDebugColor>,
+    default_collider_debug: ColliderDebug,
+    override_visibility: Query<'world, 'state, &'a ColliderDebug>,
     context: &'b RapierContext,
     gizmos: Gizmos<'world, 'state>,
 }
@@ -117,6 +142,26 @@ impl<'world, 'state, 'a, 'b> BevyLinesRenderBackend<'world, 'state, 'a, 'b> {
 
         color.map(|co: Hsla| co.to_f32_array()).unwrap_or(default)
     }
+
+    fn drawing_enabled(&self, object: DebugRenderObject) -> bool {
+        match object {
+            DebugRenderObject::Collider(h, ..) => {
+                let Some(collider) = self.context.colliders.get(h) else {
+                    return false;
+                };
+                let entity = Entity::from_bits(collider.user_data as u64);
+
+                let collider_debug =
+                    if let Ok(collider_override) = self.override_visibility.get(entity) {
+                        *collider_override
+                    } else {
+                        self.default_collider_debug
+                    };
+                collider_debug == ColliderDebug::AlwaysRender
+            }
+            _ => true,
+        }
+    }
 }
 
 impl<'world, 'state, 'a, 'b> DebugRenderBackend for BevyLinesRenderBackend<'world, 'state, 'a, 'b> {
@@ -128,6 +173,10 @@ impl<'world, 'state, 'a, 'b> DebugRenderBackend for BevyLinesRenderBackend<'worl
         b: Point<Real>,
         color: [f32; 4],
     ) {
+        if !self.drawing_enabled(object) {
+            return;
+        }
+
         let color = self.object_color(object, color);
         self.gizmos.line(
             [a.x, a.y, 0.0].into(),
@@ -144,6 +193,10 @@ impl<'world, 'state, 'a, 'b> DebugRenderBackend for BevyLinesRenderBackend<'worl
         b: Point<Real>,
         color: [f32; 4],
     ) {
+        if !self.drawing_enabled(object) {
+            return;
+        }
+
         let color = self.object_color(object, color);
         self.gizmos.line(
             [a.x, a.y, a.z].into(),
@@ -153,11 +206,12 @@ impl<'world, 'state, 'a, 'b> DebugRenderBackend for BevyLinesRenderBackend<'worl
     }
 }
 
-fn debug_render_scene(
+fn debug_render_scene<'a>(
     rapier_context: Res<RapierContext>,
     mut render_context: ResMut<DebugRenderContext>,
     gizmos: Gizmos,
-    custom_colors: Query<&ColliderDebugColor>,
+    custom_colors: Query<&'a ColliderDebugColor>,
+    override_visibility: Query<&'a ColliderDebug>,
 ) {
     if !render_context.enabled {
         return;
@@ -165,6 +219,8 @@ fn debug_render_scene(
 
     let mut backend = BevyLinesRenderBackend {
         custom_colors,
+        default_collider_debug: render_context.default_collider_debug,
+        override_visibility,
         context: &rapier_context,
         gizmos,
     };